moved step result classes to separate package
