+++ src/main/java/fr/bird/bloom/stepresults/Step5_IncludeSynonym.java
@@ -2,7 +2,7 @@
  * fr.bird.bloom.beans
  * Step5_IncludeSynonym
  */
-package fr.bird.bloom.beans;
+package fr.bird.bloom.stepresults;
 
 /**
  * fr.bird.bloom.beans
