+++ src/main/java/fr/bird/bloom/stepresults/Step6_CheckTDWG.java
@@ -2,7 +2,7 @@
  * fr.bird.bloom.beans
  * Step6_CheckTDWG
  */
-package fr.bird.bloom.beans;
+package fr.bird.bloom.stepresults;
 
 /**
  * fr.bird.bloom.beans
