+++ src/main/java/fr/bird/bloom/stepresults/Step9_EstablishmentMeans.java
@@ -2,7 +2,7 @@
  * fr.bird.bloom.beans
  * Step9_EstablishmentMeans
  */
-package fr.bird.bloom.beans;
+package fr.bird.bloom.stepresults;
 
 /**
  * fr.bird.bloom.beans
