+++ src/main/webapp/LaunchPage.html
@@ -263,10 +263,10 @@
         </div>
     </footer>
 	<script
-		src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js">
+		src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js" type="application/javascript">
     </script>
-    <script src="bootstrap/js/bootstrap.min.js"></script>
-	<script src="js/jquery.watable.js"></script>
+    <script type="application/javascript" src="bootstrap/js/bootstrap.min.js"></script>
+	<script type="application/javascript" src="js/jquery.watable.js"></script>
 	<script type="text/javascript" src="js/ReconciliationFunctions.js"></script>
 	<script type="text/javascript" src="js/functionsPrep.js"></script>
 	<script type="text/javascript" src="js/functionsMapping.js"></script>
